refactor(crypto): rename cipher module, split out secret generation
Rename cipher.rs to encryption.rs now that it is the single wire format,
and move generate_secret into its own generation.rs — random-token
minting is a distinct concern from encrypt/decrypt.
