+++ crates/nvisy-server/src/service/crypto/mod.rs
@@ -3,16 +3,18 @@
 //! This module provides encryption and decryption utilities using XChaCha20-Poly1305,
 //! a modern AEAD cipher suitable for encrypting sensitive data at rest.
 
-mod cipher;
+mod encryption;
 mod error;
+mod generation;
 mod hashing_reader;
 mod key;
 mod service;
 
-pub(crate) use cipher::{
+pub(crate) use encryption::{
     decrypt, decrypt_json, decrypt_reader, encrypt, encrypt_json, encrypt_reader,
 };
 pub use error::{CryptoError, CryptoResult};
+pub(crate) use generation::generate_secret;
 pub(crate) use hashing_reader::HashingReader;
 pub(crate) use key::EncryptionKey;
 pub use service::{CryptoConfig, CryptoService};
